pyproject.toml
[tool.poetry]
name = "jump-download"
version = "1.0.0"
description = "Code to download the JUMP Cell Painting dataset."
authors = ["Gabriel Watkinson <gabriel.watkinson@gmail.com>"]
readme = "README.md"
packages = [{include = "jump_download"}]



# Build
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Scritps
[tool.poetry.scripts]
download_metadata = "jump_download.metadata.download_jump_metadata:main"
download_load_data_files = "jump_download.load_data_files.create_total_load_data_file:main"
create_job_split = "jump_download.create_jobs.sample_load_data:sample_load_data"
download_images_from_csv = "jump_download.images.final_image_class:main"
recreate_submission_file = "jump_download.create_jobs.recreate_submission:main"



# Dependencies

[tool.poetry.dependencies]
# Base python
python = "^3.8.1"
tqdm = "^4.65.0"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"

# Data manipulation dependencies
numpy = "^1.24.3"
pandas = "^2.0.1"
tabulate = "^0.9.0"
h5py = "^3.8.0"
pyarrow = "^12.0.1"

# Image manipulation dependencies
pillow = "^9.5.0"
imagecodecs = "^2023.3.16"
tifffile = "^2023.4.12"
pylibtiff = "^0.5.1"
scikit-image = "^0.21.0"

# AWS dependencies
boto3 = "^1.26"
s3fs = "^2023.5.0"


[tool.poetry.group.dev.dependencies]
# Dev dependencies
pre-commit = "^3.1.1"
poethepoet = "^0.18.1"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
jupyter = "^1.0.0"
notebook = "^6.5.4"
ipykernel = "^6.23.1"
ipywidgets = "^8.0.6"
black = "^23.3.0"
nbqa = {extras = ["toolchain"], version = "^1.7.0"}
flake8 = "^6.0.0"
pylama = "^8.4.1"
flake8-bugbear = "^23.2.13"
flake8-comprehensions = "^3.10.1"
flake8-simplify = "^0.19.3"
pep8-naming = "^0.13.3"


# Tools configuration
[tool.isort]
profile = "black"
line_length = 100
float_to_top = true

[tool.flake8]
ignore = ['E203', 'E501', 'W503', 'N80']
max-line-length = 100
count = true
docstring-convention = 'google'

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100

[tool.pyupgrade]
py38plus = true

[tool.autoflake]
check = true
expand-star-imports = true
ignore-pass-after-docstring = true
